Convert tools to js
