+++ tools/set-metadata.js
@@ -1,7 +1,20 @@
 /* The following code is from https://github.com/ngxs/store/blob/master/tools/set-metadata.ts */
 
-import { writeFile } from 'fs';
-import { getProjects } from './utils';
+const { writeFile } = require('fs');
+const { resolve } = require('path');
+
+function getProjects() {
+  const { projects } = require('../angular.json');
+  return Object.keys(projects)
+    .map(key => {
+      const path = resolve(__dirname, '../', projects[key].root);
+      return {
+        ...projects[key],
+        path
+      };
+    })
+    .filter(project => project.root !== '');
+}
 
 async function main() {
   const ngxsJson = require('../package.json');
@@ -10,6 +23,7 @@ async function main() {
   const packages = getProjects();
   for (const pack of packages) {
     const packPath = `${pack.path}/package.json`;
+    // tslint:disable-next-line: tsr-detect-non-literal-require
     const packPackage = require(packPath);
 
     // copy all meta data from the root package.json into all packages
